Refactoring
Doing some refactoring before working on implementation details of the scoring plugin:
- Move all common types into types.go
- Move all common logic which being used by all plugins to common-plugin-helpers.go
